.history/requirements_20231119214429.txt
azure.cognitiveservices.vision.computervision
msrest.authentication import CognitiveServicesCredentials
from array import array
import os
from PIL import Image
from PIL import ImageDraw
from PIL import ImageFont
import sys
import time
import json
import streamlit as st